yt-archiver-backend/pyproject.toml
[project]
name = "yt-archiver-backend"
version = "0.1.0"
description = "Self-hosted YouTube archiving platform backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pymongo>=4.10.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "structlog>=24.0.0",
    "gunicorn>=21.2.0",
    "yt-dlp>=2023.12.30",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
